+++ src/aux-common/Files/File.ts
@@ -94,6 +94,10 @@ export interface FileTags {
     ['aux.mod']?: unknown;
     ['aux.mod.mergeTags']?: unknown;
     ['aux.creator']?: string;
+    ['aux.progressBar']?: unknown;
+    ['aux.progressBar.color']?: unknown;
+    ['aux.progressBar.backgroundColor']?: unknown;
+    ['aux.progressBar.anchor']?: unknown;
 
     // User tags
     ['aux._selection']?: string;
@@ -126,6 +130,16 @@ export interface FileTags {
     // Admin channel tags
     ['aux.connectedSessions']?: number;
     ['aux.maxSessionsAllowed']?: number;
+
+    // Admin channel task tags
+    ['aux.runningTasks']?: boolean;
+    ['aux.finishedTasks']?: boolean;
+    ['aux.task.output']?: unknown;
+    ['aux.task.error']?: unknown;
+    ['aux.task.time']?: unknown;
+    ['aux.task.shell']?: string;
+    ['aux.task.github']?: boolean;
+    ['aux.task.github.url']?: string;
 
     // Context related tags
     ['aux.context']?: string;
@@ -409,6 +423,12 @@ export const KNOWN_TAGS: string[] = [
     'aux.context.surface.minimized',
     'aux.context.surface.movable',
     'aux.context.visualize',
+    'aux.task.output',
+    'aux.task.error',
+    'aux.task.time',
+    'aux.task.shell',
+    'aux.task.github',
+    'aux.task.github.url',
     'onClick()',
     'onAnyBotClicked()',
     'onCombine()',