+++ CHANGELOG.md
@@ -8,6 +8,11 @@
 
 -   Improvements
     -   Improved the `server.shell()` command to output a bot to the `aux.finishedTasks` channel with the results of the command.
+    -   Added the ability to backup channels to Github using Gists.
+        -   You can trigger a backup by running `server.backupToGithub(token)` as an admin from the admin channel.
+        -   The `token` parameter should be replaced with a string containing a [personal access token](https://help.github.com/en/articles/creating-a-personal-access-token-for-the-command-line) from the account you want the backup to upload to.
+        -   During upload a bot will be added to the `aux.runningTasks` context with a progress bar indicating the status of the operation.
+        -   When the task is completed the bot will be moved to the `aux.finishedTasks` context and will contain tags indicating the result of the operation.
 -   Bug Fixes
     -   Removed mouse pointer change on player inventory side bars.
     -   Made the multiselect button ui consistant colors.